src/pyproject.toml
[tool.poetry]
name = "pipeline"
version = "0.1.0"
description = ""
authors = ["wyhwong <wyhwong@link.cuhk.edu.hk>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.1"
numpy = "^1.26.3"
pandas = "^2.2.0"
matplotlib = "^3.8.2"
seaborn = "^0.13.1"
opencv-python = "^4.9.0.80"
onnx = "^1.15.0"
scikit-learn = "^1.4.0"
torch = "^2.1.2"
torchvision = "^0.16.2"
tqdm = "^4.66.1"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.4"
pytest-cov = "^4.1.0"
black = "^23.12.1"
pylint = "^3.0.3"
mypy = "^1.8.0"
bandit = "^1.7.6"
radon = "^6.0.1"
isort = "^5.13.2"
pre-commit = "^3.6.0"

[tool.black]
line-length = 120

[tool.isort]
line_length = 120
include_trailing_comma = true
lines_after_imports = 2

[tool.bandit]
exclude_dirs = ["tests",]

[tool.mypy]
exclude = ["tests",]

[[tool.mypy.overrides]]
module = "*.*"
ignore_missing_imports = true

# Tutorial:
# https://radon.readthedocs.io/en/latest/commandline.html#radon-configuration-files
[tool.radon]
exclude = "test_*.py"
cc_min = "A"
cc_max = "C"
average = true
total_average = true
show_complexity = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
